requirements.txt
altgraph==0.17.4
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pywin32-ctypes==0.2.2
